requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
